+++ backend/src/main/java/com/stampcrush/backend/api/visitor/profile/request/VisitorProfilesLinkDataRequest.java
@@ -1,5 +1,6 @@
-package com.stampcrush.backend.api.visitor.profile;
+package com.stampcrush.backend.api.visitor.profile.request;
 
+import com.stampcrush.backend.application.visitor.profile.dto.VisitorProfilesLinkDataDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
